Prior work to transition to a generic approach for stable, beta and nightly releases
